view_answerを作る

+++ db/migrate/20190930032127_change_data_recruit_percent_to_input.rb
@@ -1,5 +1,6 @@
 class ChangeDataRecruitPercentToInput < ActiveRecord::Migration[5.2]
   def change
-    change_column :inputs, :recruit_percent, :integer
+    change_column :inputs, :recruit_percent, :decimal
+    
   end
 end
